BACK-END/user-service/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
pydantic[email]>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.25.0
python-dotenv>=1.0.0
redis>=5.0.0
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
cryptography>=41.0.0
tabulate>=0.9.0

# Analytics
posthog>=3.0.0

# Payment
stripe>=7.0.0